order-processing-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.2.0
eventlet==0.33.0